dal - refactored plugin settings service
